Changed color configuration to colorConfig.json and got modular colors working for ProfileButton
