+++ app/generated-colors.css
@@ -1,0 +1,11 @@
+@theme {
+--color-primary-200: #ffffff;
+--color-primary-300: #FFFFFF;
+--color-primary-400: #e2e2e2;
+--color-secondary-200: #262626;
+--color-secondary-300: #000000;
+--color-secondary-400: #000000;
+--color-accent-200: #6137ff;
+--color-accent-300: #0000FF;
+--color-accent-400: #0000e1;
+}
