+++ colorConfig.json
@@ -1,0 +1,7 @@
+{
+    "colors": {
+      "primary":   "#FFFFFF",
+      "secondary": "#000000",
+      "accent":    "#0000FF"
+    }
+  }
