+++ components/ProfileButton.tsx
@@ -31,14 +31,14 @@ export function ProfileButton({
     <Popover>
       <PopoverTrigger asChild>
         <Button variant="ghost" className="relative h-10 w-10 rounded-full">
-          <Avatar className="h-10 w-10 border-2 border-orange-300 hover:border-orange-400 transition-colors">
-            <AvatarFallback className="bg-yellow-100">
+          <Avatar className="h-10 w-10 border-2 border-[var(--color-accent-300)] hover:border-[var(--color-accent-400)] transition-colors">
+            <AvatarFallback className="bg-[var(--color-accent-200)]">
               {getInitials(fullName)}
             </AvatarFallback>
           </Avatar>
         </Button>
       </PopoverTrigger>
-      <PopoverContent className="w-80 backdrop-blur-sm bg-white/90 border border-orange-200">
+      <PopoverContent className="w-80 backdrop-blur-sm bg-[var(--color-primary-300)]/90 border border-[var(--color-accent-200)]">
         <div className="space-y-4">
           <div className="space-y-2">
             <h4 className="font-medium leading-none">
@@ -57,7 +57,7 @@ export function ProfileButton({
             <div className="flex justify-center">
               <Button
                 formAction={logout}
-                className="mt-5 bg-black text-white hover:bg-gray-800 hover:scale-105 transition-colors duration-200"
+                className="mt-5 bg-[var(--color-secondary-300)] text-[var(--color-primary-300)] hover:bg-[var(--color-secondary-200)] hover:scale-105 transition-colors duration-200"
               >
                 Logga ut
               </Button>
