+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "generate-theme": "ts-node scripts/generate-theme-css.ts",
+    "generate-theme": "node scripts/generate-theme-css.js",
     "predev": "npm run generate-theme",
     "dev": "next dev --turbopack",
     "build": "next build",
